Reimplemented Leaderboard and a proper death screen
Reimplemented the XML based leaderboard system, and added back the death screen showing progress made and current best. Player can then submit score with a name to be saved in the XML and retrieved when displaying the leaderboard after. Still TODO: Add scenemanager code to reload the scene or go back to menu when clicking buttons, function is there just unsure on how we're managing that.